Refactored styles
