+++ src/features/mode/play/panel/PlayPanel.js
@@ -10,7 +10,7 @@ import ButtonsDialogs from 'features/panel/ButtonsDialogs';
 import History from 'features/panel/History';
 import MovesTable from 'features/panel/MovesTable';
 import OpeningTable from 'features/panel/OpeningTable';
-import styles from 'styles/panel/styles';
+import styles from 'styles/panel';
 
 const PlayPanel = ({ props }) => {
   return (
