+++ src/features/mode/rav/panel/RavPanel.js
@@ -4,7 +4,7 @@ import GameMetadataTable from 'features/panel/GameMetadataTable';
 import History from 'features/panel/History';
 import RavButtons from 'features/mode/rav/panel/RavButtons';
 import RavMovesTable from 'features/mode/rav/panel/RavMovesTable';
-import styles from 'styles/panel/styles';
+import styles from 'styles/panel';
 
 const RavPanel = ({ props }) => {
   return (
