+++ src/features/mode/stockfish/panel/StockfishPanel.js
@@ -6,7 +6,7 @@ import ButtonsDialogs from 'features/panel/ButtonsDialogs';
 import History from 'features/panel/History';
 import MovesTable from 'features/panel/MovesTable';
 import OpeningTable from 'features/panel/OpeningTable';
-import styles from 'styles/panel/styles';
+import styles from 'styles/panel';
 
 const StockfishPanel = ({ props }) => {
   return (
