+++ src/styles/panel.js
